--- README.md
@@ -1,2 +1,0 @@
-# Webserv
-42cursus Webserv team project
