+++ config/our_default.conf
@@ -1,7 +1,6 @@
 
 # - host:port의 첫 번째 서버는 이 host:port의 기본값이 됩니다. 즉, 다른 서버에 속하지 않는 모든 요청에 응답합니다.
 
-
 server
 {
 	# [server] default 내용은 반드시 밑 순서를 맞추어 작성, 아니면 error
@@ -22,13 +21,48 @@ server
 		# [location] 순서는 신경쓰지 않음
 		# location에 명시되어 있는 항목들은 default값에 덮어쓰기.
 
-		error_page			404 /error_page_404.html
-		allow_methods 		GET POST DELETE
-		client_max_body_size 2048
-		root 				./html_file
-		index 				index.html upload_hoylee.html 
-		auto_index 			on
+		error_page			404 /error_page.html
+		allow_methods 		GET POST
+		client_max_body_size 2046
+		root				./html_file
+		index 				index.html root.html
+		auto_index 			off
 		return 				301 https://www.naver.com
 		cgi_extension		.bla
+	}
+
+	location /test
+	{
+		error_page			401 /access_denied.html
+		allow_methods 		POST
+		client_max_body_size 430
+		root				./test
+		index 				test.html index_test.html
+		auto_index 			off
+		return 				301 https://www.test.com
+		cgi_extension		.php
+	}
+}
+
+server
+{
+	listen					127.0.0.1 	# ip 주소가 없는 디폴트 : INADDR_ANY, port 가 없는 default : 80
+	server_name				_ 				# 없으면 빈 문자열, 하나 있으면 넣고, 두개 이상 error 문자열 구분은 ' ' 이거 뒤에 이어지면 2개임
+	root					/www 			# 현재 경로 처럼 보이지만 pwd 로 패스를 받아와서 추가하는 형식으로 넣어야 한다.
+	index					index.html
+	error_page				404 /error_page_404.html # 숫자는 여러개 나올 수 있다. 파일을 찾을 수 없을 때 parsing 에서 error 종료 // default 는 이거
+	client_max_body_size	334			# default 256, 256보다 작거나 4096보다 큰 값일 때 에러
+	auto_index 				on				# [on,off] 외 입력은 에러, 없으면 default는 off로
+
+	location /test
+	{
+		error_page			401 /access_denied.html
+		allow_methods 		POST
+		client_max_body_size 430
+		root				./test
+		index 				test.html index_test.html
+		auto_index 			off
+		return 				301 https://www.test.com
+		cgi_extension		.php
 	}
 }
