--- ours/config/default_tutorials.conf
@@ -1,70 +1,0 @@
-SOFTWARE_NAME ft_nginx
-SOFTWARE_VERSION 1.0
-HTTP_VERSION 1.1
-CGI_VERSION 1.1
-
-server {
-    server_name default
-    host 127.0.0.1
-    port 8080
-    REQUEST_URI_LIMIT_SIZE 1024
-    REQUEST_HEADER_LIMIT_SIZE 256
-    DEFAULT_ERROR_PAGE config/html/default_error.html
-    LIMIT_CLIENT_BODY_SIZE 200000000
-
-    location / {
-        root /Users/gim-eunhyul/42seoul/subject/webserv/mandatory/YoupiBanane
-        allow_method GET
-        index youpi.bad_extension
-        autoindex on
-    }
-
-    location / {
-        root /Users/gim-eunhyul/42seoul/subject/webserv/mandatory/YoupiBanane
-        allow_method POST
-        index youpi.bad_extension
-        autoindex on
-        cgi .php .bla
-    }
-
-    location /put_test {
-        root /Users/gim-eunhyul/42seoul/subject/webserv/mandatory/YoupiBanane
-        allow_method PUT
-        index index.html index.htm
-        autoindex off
-    }
-
-    location /directory {
-        root /Users/gim-eunhyul/42seoul/subject/webserv/mandatory/YoupiBanane
-        allow_method GET POST
-        index youpi.bad_extension
-        autoindex off
-        cgi .bla
-    }
-
-    location /post_body {
-        root /Users/gim-eunhyul/42seoul/subject/webserv/mandatory/YoupiBanane
-        allow_method POST
-        index youpi.bla
-        autoindex off
-        limit_client_body_size 100
-        cgi .bla
-    }
-}
-
-server {
-    server_name HelloService
-    host 127.0.0.1
-    port 8081
-    REQUEST_URI_LIMIT_SIZE 1024
-    REQUEST_HEADER_LIMIT_SIZE 256
-    DEFAULT_ERROR_PAGE config/html/simple_error.html
-    LIMIT_CLIENT_BODY_SIZE 200000000
-
-    location / {
-        root /Users/gim-eunhyul/42seoul/subject/webserv/mandatory/HelloService
-        allow_method GET
-        index index.html
-        autoindex off
-    }
-}
