--- ours/config/default_wiki.conf
@@ -1,91 +1,0 @@
-server
-{
-	listen 127.0.0.1:100	
-	#설명 : listen 127.0.0.1 ,listen 100
-	
-	error_page 300 405 304 404 /error_page_404.html
-	#설명 : 300 ~ 599까지 에러코드 사용가능
-	#설명 : DEFAULT_HTML_PATH + /error_page_404.html 경로가 열립니다.
-
-	location /
-	{
-		#설명 : 아래 기능들은 모두 사용가능합니다.
-		error_page				404 /error_page_404.html
-		allow_methods			GET POST
-		client_max_body_size	2048
-		root					./html_file
-		index					index.html
-		auto_index				on
-		#cgi_extension			.bla
-		#auth_key				hyeonskizzangzzangman:1234
-		#client_body_buffer_size 1000M #이렇게 올수도 있음.
-		#upload_path /tmp/
-#return 302 https://www.naver.com #이건 리다이렉션할떄 올수 있음.
-		#cgi_path_info : 구현안됨. 말봉님 코드에는 있는데, nginx 공홈에는 없음. 
-
-
-		######################################################################
-		### 1. error_page : server block의 error_page와 location block에 error_page가 있으면 location의 error_page만 동작합니다.
-		### 2. allow_methods : GET POST DELETE를 허용가능합니다. 해당 헤더가 없으면 GET POST DELETE는 모두 허용됩니다. 
-		### 3. client_max_body_size : 클라이언트에서 server로 용량을 제한합니다. 최대 2048MB 까지 가능합니다. 최대 2048까지 설정이 가능합니다. 
-		### 4. root : 해당 location의 root 폴더를 지정합니다. 설정하지 않으면 해당 location block의 root 는 DEFAULT_HTML_PATH 입니다.
-		### 5. index : 여러개 설정 가능합니다. 경로는 root + (url - location_path) + index 입니다.
-		### 6. auto_index : on 또는 off 설정이 가능하며 on 인 경우 index파일이 없으면 root + (url - location_path)의 리스트를 띄웁니다.
-		######################################################################
-	}
-
-	location /2////
-	{
-		root ./html_file
-		index index2.html
-		auto_index on
-	}
-
-	location /virtual//siej///iowej
-	{
-		root ./tests/www/
-		cgi_extension .bla
-	}
-
-	location v
-	{
-		allow_methods GET POST DELETE
-		root ./tests/www/tmp/
-		client_body_buffer_size 1000
-	}
-
-	location /post/tmp/
-	{
-		allow_methods POST
-		root ./test/www/tmp/
-	}
-
-	location /index/
-	{
-		root ./tests/www/
-		auth_key awesome_hyeonski
-		return 301 https://www.naver.com
-	}
-}
-
-server
-{
-	listen 127.23.12.31
-	location /
-	{
-		root ./tests/www/
-		index second.html
-	}
-}
-
-server
-{
-#	server_name google.com
-#	listen 8182 127.0.0.1
-
-	location /
-	{
-		root ./tests/www/
-		index google.html
-	}
-}
