--- ours/config/mine.conf
@@ -1,23 +1,0 @@
-server
-{
-	host				127.0.0.1
-	listen				8080
-	server_name			test_server
-	DEFAULT_ERROR_PAGE	./error.html
-
-	location /
-	{
-		root # root 폴더를 지정
-		allow_methods GET POST DELETE # GET POST DELETE 세개만 받을 것
-		index # 뭔지 모르겠지만 여러개 들어올 수 있음
-		auto_index # on off
-		limit_client_body_size # int server에서 일괄적으로 주거나 하나씩 설정할 수 있는듯
-		cgi .php
-	}
-	# 리다이렉션은 모르겠
-	# cgi도 모르겠
-}
-
-# { } # 셋이 확인되면 그냥 넘어간다.
-
-# server, location 블록은 최소 하나 이상 존재해야 함.
